Move smb tests to file_operations module
