+++ app/src/test/java/com/amaze/filemanager/ui/activities/MainActivityTest.java
@@ -46,7 +46,7 @@ import org.robolectric.shadows.ShadowStorageManager;
 
 import com.amaze.filemanager.application.AppConfig;
 import com.amaze.filemanager.shadows.ShadowMultiDex;
-import com.amaze.filemanager.shadows.jcifs.smb.ShadowSmbFile;
+import com.amaze.filemanager.file_operations.shadows.jcifs.smb.ShadowSmbFile;
 import com.amaze.filemanager.test.ShadowCryptUtil;
 import com.amaze.filemanager.test.TestUtils;
 import com.amaze.filemanager.utils.SmbUtil;
