+++ file_operations/src/test/java/com/amaze/filemanager/file_operations/filesystem/smbstreamer/StreamSourceTest.java
@@ -18,7 +18,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package com.amaze.filemanager.filesystem.smbstreamer;
+package com.amaze.filemanager.file_operations.filesystem.smbstreamer;
 
 import static android.os.Build.VERSION_CODES.JELLY_BEAN;
 import static android.os.Build.VERSION_CODES.KITKAT;
@@ -40,8 +40,8 @@ import org.robolectric.annotation.Config;
 import org.robolectric.shadow.api.Shadow;
 
 import com.amaze.filemanager.file_operations.filesystem.smbstreamer.StreamSource;
-import com.amaze.filemanager.shadows.ShadowMultiDex;
-import com.amaze.filemanager.shadows.jcifs.smb.ShadowSmbFile;
+import com.amaze.filemanager.file_operations.shadows.ShadowMultiDex;
+import com.amaze.filemanager.file_operations.shadows.jcifs.smb.ShadowSmbFile;
 
 import android.os.Environment;
